+++ app/src/main/java/net/typeblog/shelter/ui/SettingsFragment.java
@@ -31,7 +31,6 @@ public class SettingsFragment extends PreferenceFragmentCompat implements Prefer
     private static final String SETTINGS_BUG_REPORT = "settings_bug_report";
     private static final String SETTINGS_PATREON = "settings_patreon";
     private static final String SETTINGS_CROSS_PROFILE_FILE_CHOOSER = "settings_cross_profile_file_chooser";
-    private static final String SETTINGS_CAMERA_PROXY = "settings_camera_proxy";
     private static final String SETTINGS_BLOCK_CONTACTS_SEARCHING = "settings_block_contacts_searching";
     private static final String SETTINGS_AUTO_FREEZE_SERVICE = "settings_auto_freeze_service";
     private static final String SETTINGS_AUTO_FREEZE_DELAY = "settings_auto_freeze_delay";
@@ -41,7 +40,6 @@ public class SettingsFragment extends PreferenceFragmentCompat implements Prefer
     private IShelterService mServiceWork = null;
 
     private CheckBoxPreference mPrefCrossProfileFileChooser = null;
-    private CheckBoxPreference mPrefCameraProxy = null;
     private CheckBoxPreference mPrefBlockContactsSearching = null;
     private CheckBoxPreference mPrefAutoFreezeService = null;
     private CheckBoxPreference mPrefSkipForeground = null;
@@ -77,9 +75,6 @@ public class SettingsFragment extends PreferenceFragmentCompat implements Prefer
         mPrefCrossProfileFileChooser = (CheckBoxPreference) findPreference(SETTINGS_CROSS_PROFILE_FILE_CHOOSER);
         mPrefCrossProfileFileChooser.setChecked(mManager.getCrossProfileFileChooserEnabled());
         mPrefCrossProfileFileChooser.setOnPreferenceChangeListener(this);
-        mPrefCameraProxy = (CheckBoxPreference) findPreference(SETTINGS_CAMERA_PROXY);
-        mPrefCameraProxy.setChecked(mManager.getCameraProxyEnabled());
-        mPrefCameraProxy.setOnPreferenceChangeListener(this);
         mPrefBlockContactsSearching = (CheckBoxPreference) findPreference(SETTINGS_BLOCK_CONTACTS_SEARCHING);
         mPrefBlockContactsSearching.setChecked(mManager.getBlockContactsSearchingEnabled());
         mPrefBlockContactsSearching.setOnPreferenceChangeListener(this);
@@ -99,11 +94,6 @@ public class SettingsFragment extends PreferenceFragmentCompat implements Prefer
         // Supported on R and beyond
         if (Build.VERSION.SDK_INT == Build.VERSION_CODES.Q) {
             mPrefCrossProfileFileChooser.setEnabled(false);
-        }
-
-        // Disable fake camera on R because third-party camera activities are now unsupported
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-            mPrefCameraProxy.setEnabled(false);
         }
 
         // Disable FileShuttle for Android Go
@@ -186,9 +176,6 @@ public class SettingsFragment extends PreferenceFragmentCompat implements Prefer
             }
 
             mManager.setCrossProfileFileChooserEnabled(true);
-            return true;
-        } else if (preference == mPrefCameraProxy) {
-            mManager.setCameraProxyEnabled(((boolean) newState));
             return true;
         } else if (preference == mPrefBlockContactsSearching) {
             mManager.setBlockContactsSearchingEnabled((boolean) newState);